WIP queue metrics

+++ commonspace/sync/tryaddqueue.go
@@ -3,6 +3,7 @@ package sync
 import (
 	"context"
 	"sync"
+	"sync/atomic"
 
 	"github.com/cheggaaa/mb/v3"
 	"go.uber.org/zap"
@@ -11,6 +12,7 @@ import (
 type entry struct {
 	call     func()
 	onRemove func()
+	cnt      uint64
 }
 
 func newTryAddQueue(workers, maxSize int) *tryAddQueue {
@@ -29,52 +31,76 @@ type tryAddQueue struct {
 	ctx     context.Context
 	cancel  context.CancelFunc
 	workers int
-
+	cnt     atomic.Uint64
 	entries map[string]entry
 	batch   *mb.MB[string]
 	mx      sync.Mutex
 }
 
 func (rp *tryAddQueue) Replace(id string, call, remove func()) {
+	curCnt := rp.cnt.Load()
+	rp.cnt.Add(1)
 	rp.mx.Lock()
 	if prevEntry, ok := rp.entries[id]; ok {
 		rp.entries[id] = entry{
 			call:     call,
 			onRemove: remove,
+			cnt:      curCnt,
 		}
 		rp.mx.Unlock()
 		prevEntry.onRemove()
 		return
 	}
-	rp.entries[id] = entry{
+	ent := entry{
 		call:     call,
 		onRemove: remove,
+		cnt:      curCnt,
 	}
+	rp.entries[id] = ent
 	rp.mx.Unlock()
 	err := rp.batch.TryAdd(id)
 	if err != nil {
 		rp.mx.Lock()
 		curEntry := rp.entries[id]
-		delete(rp.entries, id)
+		if curEntry.cnt == curCnt {
+			delete(rp.entries, id)
+		}
 		rp.mx.Unlock()
-		if curEntry.onRemove != nil {
-			curEntry.onRemove()
+		if ent.onRemove != nil {
+			ent.onRemove()
 		}
 	}
 }
 
 func (rp *tryAddQueue) TryAdd(id string, call, remove func()) bool {
+	curCnt := rp.cnt.Load()
+	rp.cnt.Add(1)
 	rp.mx.Lock()
 	if _, ok := rp.entries[id]; ok {
 		rp.mx.Unlock()
+		if remove != nil {
+			remove()
+		}
 		return false
 	}
-	rp.entries[id] = entry{
+	ent := entry{
 		call:     call,
 		onRemove: remove,
+		cnt:      curCnt,
 	}
+	rp.entries[id] = ent
 	rp.mx.Unlock()
-	if err := rp.batch.TryAdd(id); err != nil {
+	err := rp.batch.TryAdd(id)
+	if err != nil {
+		rp.mx.Lock()
+		curEntry := rp.entries[id]
+		if curEntry.cnt == curCnt {
+			delete(rp.entries, id)
+		}
+		rp.mx.Unlock()
+		if ent.onRemove != nil {
+			ent.onRemove()
+		}
 		return false
 	}
 	return true
@@ -99,7 +125,9 @@ func (rp *tryAddQueue) sendLoop() {
 		rp.mx.Unlock()
 		if curEntry.call != nil {
 			curEntry.call()
-			curEntry.onRemove()
+			if curEntry.onRemove != nil {
+				curEntry.onRemove()
+			}
 		}
 	}
 }
